178 Fixing configuration
