+++ resilience4j-spring-boot/build.gradle
@@ -3,6 +3,7 @@ dependencies {
     compile ( libraries.spring_boot_actuator )
     compile ( libraries.spring_boot_web )
     compile ( libraries.spring_reactor )
+    compile ( libraries.hibernate_validator )
     compileOnly ( libraries.spring_boot_config_processor )
     compile project(':resilience4j-annotations')
     compile project(':resilience4j-spring')
