+++ resilience4j-spring-boot2/src/main/java/io/github/resilience4j/circuitbreaker/autoconfigure/CircuitBreakerPropertiesAutoConfigured.java
@@ -17,7 +17,9 @@ package io.github.resilience4j.circuitbreaker.autoconfigure;
 
 import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.validation.annotation.Validated;
 
+@Validated
 @ConfigurationProperties(prefix = "resilience4j.circuitbreaker")
 public class CircuitBreakerPropertiesAutoConfigured extends CircuitBreakerProperties {
